+++ django_iso_3166/models/region.py
@@ -1,7 +1,6 @@
 from django.db import models
 
 class Region(models.Model):
-    id = models.BigAutoField(primary_key=True)
     name = models.CharField(max_length=100)
     translations = models.TextField(blank=True, null=True)  # Changed from JSONField to TextField to match SQL
     created_at = models.DateTimeField(blank=True, null=True)
@@ -14,5 +13,8 @@ class Region(models.Model):
         return self.name
 
     class Meta:
+        managed = False
+        db_table = 'regions'
+
         verbose_name = "Region"
         verbose_name_plural = "Regions"